requirements.txt
asgiref==3.9.1
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
gunicorn==23.0.0
packaging==25.0
pillow==11.3.0
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2
